requirements.txt
PyYAML==3.11
selenium==2.46.0
wsgiref==0.1.2
